requirements.txt
#sentence-transformers==0.3.2
django==3.2.0
gunicorn==20.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==v1.7.1
mysqlclient==2.2.4
pymysql==1.1.1